requirements.txt
Django==1.10
django-chartjs==1.0
djangorestframework==3.4.6
gunicorn==19.6.0
six==1.10.0
